fix execve trace
- the argv wasn't printed correctly because it wasn't stored as a long
- print the trace of execve like strace does

+++ syscall_data.h
@@ -213,7 +213,7 @@ typedef clone_arg_s *clone_arg_t;
 typedef struct execve_arg_s {
   int ret;
   long ptr_filename;
-  int ptr_argv;
+  long ptr_argv;
 } execve_arg_s;
 typedef execve_arg_s *execve_arg_t;
